Sessions handlers
That's the first module of library folder. "Sessions" contains three differents functions.

The "get" function, return an array with session information, this array contains objects with two properties. The first property is named "user", this value must take a user name of type "string".

The secondary property is named 'time', this value must take milliseconds represented by the user's connection time. This value is of type "number".

The "start" function, take a string username as a parameter. When called, this function pushes the username into a list of sessions with the current time. (available with the function "get"). If this function is not called with the same username within fifteen minutes, those sessions automatically expire.

The last function "stop", take in parameters an username as a "string", when called, this function stop the session of username given.