Users get more properties
More properties are added in users array.

The third property is named 'score', this value must take the score of an user. In case of competive relation with others users, like games or/and comparisons. This value is of type 'number'.

The fourth property is named 'location', this value must take the location where's the user, like last requested adress or/and real location. This value is of type 'string'.

Last property is named 'items', this value must take one or several items like shopping cart or/and object represented in your application. This value is of type 'string'.