Score, Rated and Rank to users properties.
In case of competive relation with others users, like games or/and comparisons. Scores can be of any lenght of you want, rated values evaluate scores by hundred, and influence his position in the rank. Rank position is represented by numbers, to be as simple and adapted as possible to the architecture of your project.